requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
appdirs==1.4.4
APScheduler==3.9.1
asgiref==3.5.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
fastapi==0.70.0
frozenlist==1.3.0
google==3.0.0
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.2
jsonpickle==2.1.0
kafka-python==2.0.2
lxml==4.8.0
multidict==6.0.2
multitasking==0.0.10
numpy==1.21.5
oauthlib==3.2.0
pandas==1.3.5
protobuf==3.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2
SQLAlchemy==1.4.26
starlette==0.16.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.15.0
yarl==1.7.2
yfinance==0.1.70
zipp==3.7.0
